pyproject.toml
[project]
name = "sploitgpt"
version = "0.1.0"
description = "Autonomous AI Penetration Testing Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "SploitGPT Contributors"}
]
keywords = ["pentesting", "security", "ai", "llm", "metasploit", "kali"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "textual>=0.89.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "msgpack>=1.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "litellm>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
training = [
    "unsloth>=2024.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "torch>=2.2.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
]

[project.scripts]
sploitgpt = "sploitgpt.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
